fill changelog and fix several issues

+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/request/RequestId.kt
@@ -7,7 +7,7 @@ import kotlin.random.Random
 @Serializable
 @JvmInline
 value class RequestId(
-    val float: Int
+    val int: Int
 ) {
     companion object {
         fun random() = RequestId(Random.nextInt())
